requirements.txt
fastapi
uvicorn
httpx
python-dotenv
mcp
streamlit
openai
